// src/types.ts
export interface IIngredient {
    __v: number;
    _id: string;
    calories: number;
    carbohydrates: number;
    fat: number;
    image: string;
    image_large: string;
    image_mobile: string;
    name: string;
    price: number;
    proteins: number;
    type: 'bun' | 'main' | 'sauce';
    count?: number;
}